refactor: simplify command modules
- rename the `args` module to `commands`
- Move all the each sub commands into command/xxx.rs
- rename modules to simpler names (completions, schema, pull_request)
- rename init to init_github as there may be an eventual init command
  that is just for release-plz.yml
- push code down into modules by adding run() methods on each command
- move command help text into each command struct
- add aliases for backward compatibility of command names
- reorder commands based on a natural ordering which they'd be likely
  encountered (init first, admin-ish commands last)
- fix up help text generally
- Move CliArgs to main
