+++ crates/release_plz/Cargo.toml
@@ -28,7 +28,7 @@ anyhow.workspace = true
 base64.workspace = true
 cargo_metadata.workspace = true
 chrono = { workspace = true, features = ["clock"] }
-clap = { workspace = true, features = ["derive", "env"] }
+clap = { workspace = true, features = ["derive", "env", "wrap_help"] }
 clap_complete.workspace = true
 dirs.workspace = true
 fs-err.workspace = true
