+++ crates/release_plz/src/commands/manifest.rs
@@ -2,7 +2,8 @@ use std::path::Path;
 
 use anyhow::Context as _;
 use cargo_metadata::camino::{Utf8Path, Utf8PathBuf};
-use release_plz_core::fs_utils::to_utf8_path;
+use cargo_utils::CARGO_TOML;
+use release_plz_core::fs_utils::{current_directory, to_utf8_path};
 
 /// Command that acts on a manifest.
 pub trait ManifestCommand {
@@ -13,7 +14,7 @@ pub trait ManifestCommand {
     }
 
     fn manifest_path(&self) -> Utf8PathBuf {
-        super::local_manifest(self.optional_manifest_path())
+        local_manifest(self.optional_manifest_path())
     }
 
     fn cargo_metadata(&self) -> anyhow::Result<cargo_metadata::Metadata> {
@@ -27,5 +28,12 @@ pub trait ManifestCommand {
                 anyhow::anyhow!(e)
             }
         }).with_context(|| format!("Failed to read metadata from manifest at {manifest}"))
+    }
+}
+
+fn local_manifest(manifest_path: Option<&Utf8Path>) -> Utf8PathBuf {
+    match manifest_path {
+        Some(manifest) => manifest.to_path_buf(),
+        None => current_directory().unwrap().join(CARGO_TOML),
     }
 }
